+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## [0.0.2] - 2021.06.12
+* Add 'persistent' boolean to show() function in order to not dismiss the notification over time.
+* Add animations set to customization, as well as custom animations.
+
 ## [0.0.1] - 2021.05.30
 * Initial version